+++ src/v1_6/types/charging_schedule.rs
@@ -8,7 +8,7 @@ use super::{ChargingRateUnitType, ChargingSchedulePeriod};
 pub struct ChargingSchedule {
     /// Optional. Duration of the charging schedule in seconds. If the duration is left empty, the last period will continue indefinitely or until end of the transaction in case startSchedule is absent.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub duration: Option<i64>,
+    pub duration: Option<i32>,
     /// Optional. Starting point of an absolute schedule. If absent the schedule will be relative to start of charging.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub start_schedule: Option<DateTime<Utc>>,
