+++ src/v1_6/types/charging_schedule_period.rs
@@ -3,10 +3,10 @@
 #[serde(rename_all = "camelCase")]
 pub struct ChargingSchedulePeriod {
     /// Required. Start of the period, in seconds from the start of schedule. The value of StartPeriod also defines the stop time of the previous period.
-    pub start_period: i64,
+    pub start_period: i32,
     /// Required. Charging rate limit during the schedule period, in the applicable chargingRateUnit, for example in Amperes or Watts. Accepts at most one digit fraction (e.g. 8.1).
     pub limit: f64,
     /// Optional. The number of phases that can be used for charging. If a number of phases is needed, numberPhases=3 will be assumed unless another number is given.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub number_phases: Option<i64>,
+    pub number_phases: Option<i32>,
 }
