+++ src/v2_0_1/datatypes/charging_schedule_type.rs
@@ -12,13 +12,13 @@ use crate::v2_0_1::enumerations::charging_rate_unit_enum_type::ChargingRateUnitE
 #[serde(rename_all = "camelCase")]
 pub struct ChargingScheduleType {
     /// Required. Identifies the ChargingSchedule.
-    pub id: i64,
+    pub id: i32,
     /// Optional. Starting point of an absolute schedule. If absent the schedule will be relative to start of charging
     #[serde(skip_serializing_if = "Option::is_none")]
     pub start_schedule: Option<DateTime<Utc>>,
     /// Optional. Duration of the charging schedule in seconds. If the duration is left empty, the last period will continue indefinitely or until end of the transaction if chargingProfilePurpose = TxProfile
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub duration: Option<i64>,
+    pub duration: Option<i32>,
     /// Required. The unit of measure Limit is expressed in.
     pub charging_rate_unit: ChargingRateUnitEnumType,
     /// Optional. Minimum charging rate supported by the EV. The unit of measure is defined by the chargingRateUnit. This parameter is intended to be used by a local smart charging algorithm to optimize the power allocation for in the case a charging process is inefficient at lower charging rates. Accepts at most one digit fraction (e.g. 8.1)
