+++ src/v2_0_1/datatypes/event_data_type.rs
@@ -11,11 +11,11 @@ use crate::v2_0_1::enumerations::event_trigger_enum_type::EventTriggerEnumType;
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct EventDataType {
-    pub event_id: i64,
+    pub event_id: i32,
     pub timestamp: DateTime<Utc>,
     pub trigger: EventTriggerEnumType,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cause: Option<i64>,
+    pub cause: Option<i32>,
     pub actual_value: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tech_code: Option<String>,
@@ -26,7 +26,7 @@ pub struct EventDataType {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub transaction_id: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub variable_monitoring_id: Option<i64>,
+    pub variable_monitoring_id: Option<i32>,
     pub event_notification_type: EventNotificationEnumType,
     pub component: ComponentType,
     pub variable: VariableType,
