+++ src/v2_0_1/datatypes/id_token_info_type.rs
@@ -14,11 +14,11 @@ pub struct IdTokenInfoType {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cache_expiry_date_time: Option<DateTime<Utc>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub charging_priority: Option<i64>,
+    pub charging_priority: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub language1: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub evse_id: Option<Vec<i64>>,
+    pub evse_id: Option<Vec<i32>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub language2: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
