+++ src/v2_0_1/messages/meter_values.rs
@@ -5,7 +5,7 @@ use crate::v2_0_1::datatypes::meter_value_type::MeterValueType;
 #[serde(rename_all = "camelCase")]
 pub struct MeterValuesRequest {
     /// This contains a number (>0) designating an EVSE of the Charging Station. ‘0’ (zero) is used to designate the main power meter.
-    pub evse_id: i64,
+    pub evse_id: i32,
     /// The sampled meter values with timestamps.
     pub meter_value: Vec<MeterValueType>,
 }
