+++ src/v2_0_1/messages/set_charging_profile.rs
@@ -6,7 +6,7 @@ use crate::v2_0_1::enumerations::charging_profile_status_enum_type::ChargingProf
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct SetChargingProfileRequest {
-    pub evse_id: i64,
+    pub evse_id: i32,
     pub charging_profile: ChargingProfileType,
 }
